+++ custom_components/adtpulse/__init__.py
@@ -21,6 +21,7 @@ from homeassistant.const import (
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryAuthFailed, ConfigEntryNotReady
 from homeassistant.helpers.config_entry_flow import FlowResult
+from homeassistant.helpers.config_validation import config_entry_only_config_schema
 from homeassistant.helpers.typing import ConfigType
 from pyadtpulse import PyADTPulse
 from pyadtpulse.const import (
@@ -41,6 +42,8 @@ from .coordinator import ADTPulseDataUpdateCoordinator
 LOG = getLogger(__name__)
 
 SUPPORTED_PLATFORMS = ["alarm_control_panel", "binary_sensor"]
+
+CONFIG_SCHEMA = config_entry_only_config_schema(ADTPULSE_DOMAIN)
 
 
 async def async_setup(
@@ -133,6 +136,11 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         hass.bus.async_listen_once(EVENT_HOMEASSISTANT_STOP, coordinator.stop)
     )
     entry.async_on_unload(entry.add_update_listener(options_listener))
+
+    async def handle_relogin(dummy: str) -> None:  # pylint: disable=unused-argument
+        await service.async_quick_relogin()
+
+    hass.services.async_register(ADTPULSE_DOMAIN, "quick_relogin", handle_relogin)
     return True